// src/api/shift/enums/ShiftEnum.ts
export enum FilterDateType {
  CHECK_IN = 'CHECK_IN',
  CHECK_OUT = 'CHECK_OUT',
}

export enum SearchTextType {
  NICK_NAME = 'NICK_NAME',
  WALLET_ADDRESS = 'WALLET_ADDRESS',
}

export enum ShiftOrderDirection {
  ASC = 'ASC',
  DESC = 'DESC',
}

export enum ShiftOrderField {
  checkIn = 'checkIn',
  checkOut = 'checkOut',
  orderAmount = 'orderAmount',
  duringTime = 'duringTime',
}